rename convert/read.py to convert/api.py and move merge function in
