+++ echoregions/__init__.py
@@ -1,12 +1,10 @@
+from .convert.api import read_ecs, read_evl, read_evr
 from .convert.ecs_parser import CalibrationParser
-from .convert.merge import merge
-from .convert.read import read_ecs, read_evl, read_evr
 from .convert.utils import parse_simrad_fname_time, parse_time
 from .formats import Lines, Regions2D
 
 __all__ = [
     "CalibrationParser",
-    "merge",
     "read_ecs",
     "read_evl",
     "read_evr",
