+++ echoregions/convert/api.py
@@ -1,6 +1,6 @@
-from ..convert.ecs_parser import CalibrationParser
 from ..formats.lines import Lines
 from ..formats.regions2d import Regions2D
+from .ecs_parser import CalibrationParser
 
 
 def read_evr(
@@ -83,3 +83,51 @@ def read_ecs(filepath, ignore_comments=True) -> "CalibrationParser":
     return CalibrationParser(
         input_file=filepath, parse=True, ignore_comments=ignore_comments
     )
+
+
+def merge(objects, reindex_ids=False):
+    """Merge echoregion objects.
+    Currently only supports merging Regions2D objects.
+
+    Parameters
+    ----------
+    objects : list
+        a list collection of Regions2D objects
+
+    Returns
+    -------
+    combined : Regions2D
+        A Regions2D object with region ids prepended by the EVR original filename.
+    """
+    if not isinstance(objects, list) or not all(
+        isinstance(o, Regions2D) for o in objects
+    ):
+        return ValueError("`merge` takes a list of Regions2D objects")
+
+    merged_idx = []
+    merged_data = []
+    for regions in objects:
+        if not regions.data:
+            raise ValueError("EVR file has not been parsed. Call `parse_file` first.")
+        merged_data += regions.data["regions"].values()
+        if reindex_ids:
+            merged_idx += list(range(len(regions.data["regions"])))
+        else:
+            merged_idx += [
+                f"{regions.data['metadata']['file_name']}_{r}"
+                for r in regions.region_ids
+            ]
+    # Attach region information to region ids
+    merged = dict(zip(merged_idx, merged_data))
+    # Create new Regions2D object
+    merged_obj = Regions2D()
+    # Populate metadata
+    merged_obj.data["metadata"] = objects[0].data["metadata"]
+    # Combine metadata of all Regions2D objects
+    for field in objects[0].data["metadata"].keys():
+        merged_obj.data["metadata"][field] = [
+            o.data["metadata"][field] for o in objects
+        ]
+    # Set region data
+    merged_obj.data["regions"] = merged
+    return merged_obj
